Raise version to Gradle 5.0
Change expectations of deprecation warning tests to not assert the version.
There is not need to test the exact wording of the message, that just makes
the test brittle. Good wording is ensured through code review. The test is
just to ensure that the message is logged at the right time.